Feature/436 manage security on dashboards

+++ src/main/java/io/suricate/monitoring/controllers/api/ProjectController.java
@@ -27,6 +27,7 @@ import io.suricate.monitoring.model.dto.websocket.WebsocketClient;
 import io.suricate.monitoring.model.entity.project.Project;
 import io.suricate.monitoring.model.entity.project.ProjectWidget;
 import io.suricate.monitoring.model.entity.user.User;
+import io.suricate.monitoring.model.enums.ApiErrorEnum;
 import io.suricate.monitoring.service.api.ConfigurationService;
 import io.suricate.monitoring.service.api.ProjectService;
 import io.suricate.monitoring.service.api.ProjectWidgetService;
@@ -35,6 +36,7 @@ import io.suricate.monitoring.service.mapper.ProjectMapper;
 import io.suricate.monitoring.service.mapper.ProjectWidgetMapper;
 import io.suricate.monitoring.service.mapper.UserMapper;
 import io.suricate.monitoring.service.webSocket.DashboardWebSocketService;
+import io.suricate.monitoring.utils.exception.ApiException;
 import io.suricate.monitoring.utils.exception.NoContentException;
 import io.suricate.monitoring.utils.exception.ObjectNotFoundException;
 import io.swagger.annotations.*;
@@ -43,6 +45,7 @@ import org.springframework.context.annotation.Lazy;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
@@ -232,6 +235,7 @@ public class ProjectController {
     /**
      * Update an existing project
      *
+     * @param authentication    The connected user
      * @param projectToken      The project token to update
      * @param projectRequestDto The informations to update
      * @return The project updated
@@ -245,7 +249,8 @@ public class ProjectController {
     })
     @PutMapping(value = "/v1/projects/{projectToken}")
     @PreAuthorize("hasRole('ROLE_USER')")
-    public ResponseEntity<Void> updateProject(@ApiParam(name = "projectToken", value = "The project token", required = true)
+    public ResponseEntity<Void> updateProject(@ApiIgnore OAuth2Authentication authentication,
+                                              @ApiParam(name = "projectToken", value = "The project token", required = true)
                                               @PathVariable("projectToken") String projectToken,
                                               @ApiParam(name = "projectResponseDto", value = "The project information", required = true)
                                               @RequestBody ProjectRequestDto projectRequestDto) {
@@ -254,8 +259,13 @@ public class ProjectController {
             throw new ObjectNotFoundException(Project.class, projectToken);
         }
 
+        Project project = projectOptional.get();
+        if (!projectService.isConnectedUserCanAccessToProject(project, authentication.getUserAuthentication())) {
+            throw new ApiException("The user is not allowed to modify this resource", ApiErrorEnum.NOT_AUTHORIZED);
+        }
+
         projectService.updateProject(
-            projectOptional.get(),
+            project,
             projectRequestDto.getName(),
             projectRequestDto.getWidgetHeight(),
             projectRequestDto.getMaxColumn(),
@@ -268,7 +278,8 @@ public class ProjectController {
     /**
      * Method that delete a project
      *
-     * @param projectToken The project token to delete
+     * @param authentication The connected user
+     * @param projectToken   The project token to delete
      * @return The project deleted
      */
     @ApiOperation(value = "Delete a project by the project token")
@@ -281,13 +292,18 @@ public class ProjectController {
     @DeleteMapping(value = "/v1/projects/{projectToken}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @Transactional
-    public ResponseEntity<Void> deleteOneById(@ApiParam(name = "projectToken", value = "The project token", required = true)
+    public ResponseEntity<Void> deleteOneById(@ApiIgnore OAuth2Authentication authentication,
+                                              @ApiParam(name = "projectToken", value = "The project token", required = true)
                                               @PathVariable("projectToken") String projectToken) {
         Optional<Project> projectOptional = projectService.getOneByToken(projectToken);
         if (!projectOptional.isPresent()) {
             throw new ObjectNotFoundException(Project.class, projectToken);
         }
 
+        Project project = projectOptional.get();
+        if (!projectService.isConnectedUserCanAccessToProject(project, authentication.getUserAuthentication())) {
+            throw new ApiException("The user is not allowed to delete this resource", ApiErrorEnum.NOT_AUTHORIZED);
+        }
         projectService.deleteProject(projectOptional.get());
         return ResponseEntity.noContent().build();
     }
@@ -295,6 +311,7 @@ public class ProjectController {
     /**
      * Update the list of widget positions for a project
      *
+     * @param authentication                   The connected user
      * @param projectToken                     The project token to update
      * @param projectWidgetPositionRequestDtos The list of project widget positions
      * @return The project updated
@@ -308,14 +325,19 @@ public class ProjectController {
     })
     @PutMapping(value = "/v1/projects/{projectToken}/projectWidgetPositions")
     @PreAuthorize("hasRole('ROLE_USER')")
-    public ResponseEntity<Void> updateProjectWidgetsPositionForProject(@ApiParam(name = "projectToken", value = "The project token", required = true)
+    public ResponseEntity<Void> updateProjectWidgetsPositionForProject(@ApiIgnore OAuth2Authentication authentication,
+                                                                       @ApiParam(name = "projectToken", value = "The project token", required = true)
                                                                        @PathVariable("projectToken") String projectToken,
                                                                        @ApiParam(name = "projectWidgetPositionRequestDtos", value = "The list of the new positions", required = true)
                                                                        @RequestBody List<ProjectWidgetPositionRequestDto> projectWidgetPositionRequestDtos) {
         Optional<Project> projectOptional = projectService.getOneByToken(projectToken);
-
         if (!projectOptional.isPresent()) {
             throw new ObjectNotFoundException(Project.class, projectToken);
+        }
+
+        Project project = projectOptional.get();
+        if (!projectService.isConnectedUserCanAccessToProject(project, authentication.getUserAuthentication())) {
+            throw new ApiException("The user is not allowed to modify this resource", ApiErrorEnum.NOT_AUTHORIZED);
         }
 
         projectWidgetService.updateWidgetPositionByProject(projectOptional.get(), projectWidgetPositionRequestDtos);
@@ -356,6 +378,7 @@ public class ProjectController {
     /**
      * Add widget into the dashboard
      *
+     * @param authentication          The connected user
      * @param projectToken            The project token
      * @param projectWidgetRequestDto The projectWidget to add
      * @return The project
@@ -370,12 +393,19 @@ public class ProjectController {
     @PostMapping(value = "/v1/projects/{projectToken}/projectWidgets")
     @PreAuthorize("hasRole('ROLE_USER')")
     @Transactional
-    public ResponseEntity<ProjectWidgetResponseDto> addProjectWidgetToProject(@ApiParam(name = "projectToken", value = "The project token", required = true)
+    public ResponseEntity<ProjectWidgetResponseDto> addProjectWidgetToProject(@ApiIgnore OAuth2Authentication authentication,
+                                                                              @ApiParam(name = "projectToken", value = "The project token", required = true)
                                                                               @PathVariable("projectToken") String projectToken,
                                                                               @ApiParam(name = "projectWidgetDto", value = "The project widget info's", required = true)
                                                                               @RequestBody ProjectWidgetRequestDto projectWidgetRequestDto) {
-        if (!this.projectService.isProjectExists(projectToken)) {
+        Optional<Project> projectOptional = projectService.getOneByToken(projectToken);
+        if (!projectOptional.isPresent()) {
             throw new ObjectNotFoundException(Project.class, projectToken);
+        }
+
+        Project project = projectOptional.get();
+        if (!projectService.isConnectedUserCanAccessToProject(project, authentication)) {
+            throw new ApiException("The user is not allowed to modify this resource", ApiErrorEnum.NOT_AUTHORIZED);
         }
 
         ProjectWidget projectWidget = projectWidgetMapper.toNewProjectWidget(projectWidgetRequestDto, projectToken);
@@ -423,8 +453,9 @@ public class ProjectController {
     /**
      * Add a user to a project
      *
-     * @param projectToken Token of the project
-     * @param usernameMap  Username of the user to add
+     * @param authentication The connected user
+     * @param projectToken   Token of the project
+     * @param usernameMap    Username of the user to add
      * @return The project
      */
     @ApiOperation(value = "Add a user to a project")
@@ -437,18 +468,25 @@ public class ProjectController {
     })
     @PostMapping(value = "/v1/projects/{projectToken}/users")
     @PreAuthorize("hasRole('ROLE_USER')")
-    public ResponseEntity<Void> addUserToProject(@ApiParam(name = "projectToken", value = "The project token", required = true)
+    public ResponseEntity<Void> addUserToProject(@ApiIgnore OAuth2Authentication authentication,
+                                                 @ApiParam(name = "projectToken", value = "The project token", required = true)
                                                  @PathVariable("projectToken") String projectToken,
                                                  @ApiParam(name = "usernameMap", value = "A map with the username", required = true)
                                                  @RequestBody Map<String, String> usernameMap) {
-        Optional<User> userOptional = userService.getOneByUsername(usernameMap.get("username"));
-        if (!userOptional.isPresent()) {
-            throw new ObjectNotFoundException(User.class, usernameMap.get("username"));
-        }
 
         Optional<Project> projectOptional = projectService.getOneByToken(projectToken);
         if (!projectOptional.isPresent()) {
             throw new ObjectNotFoundException(Project.class, projectToken);
+        }
+
+        Project project = projectOptional.get();
+        if (!projectService.isConnectedUserCanAccessToProject(project, authentication.getUserAuthentication())) {
+            throw new ApiException("The user is not allowed to modify this resource", ApiErrorEnum.NOT_AUTHORIZED);
+        }
+
+        Optional<User> userOptional = userService.getOneByUsername(usernameMap.get("username"));
+        if (!userOptional.isPresent()) {
+            throw new ObjectNotFoundException(User.class, usernameMap.get("username"));
         }
 
         projectService.addUserToProject(userOptional.get(), projectOptional.get());
@@ -458,8 +496,9 @@ public class ProjectController {
     /**
      * Delete a user from a dashboard
      *
-     * @param projectToken The project/dashboard token
-     * @param userId       The user id to delete
+     * @param authentication The connected user
+     * @param projectToken   The project/dashboard token
+     * @param userId         The user id to delete
      * @return The project
      */
     @ApiOperation(value = "Delete a user from a project")
@@ -472,18 +511,24 @@ public class ProjectController {
     })
     @DeleteMapping(value = "/v1/projects/{projectToken}/users/{userId}")
     @PreAuthorize("hasRole('ROLE_USER')")
-    public ResponseEntity<Void> deleteUserToProject(@ApiParam(name = "projectToken", value = "The project token", required = true)
+    public ResponseEntity<Void> deleteUserToProject(@ApiIgnore OAuth2Authentication authentication,
+                                                    @ApiParam(name = "projectToken", value = "The project token", required = true)
                                                     @PathVariable("projectToken") String projectToken,
                                                     @ApiParam(name = "userId", value = "The user id", required = true)
                                                     @PathVariable("userId") Long userId) {
-        Optional<User> userOptional = userService.getOne(userId);
-        if (!userOptional.isPresent()) {
-            throw new ObjectNotFoundException(User.class, userId);
-        }
-
         Optional<Project> projectOptional = projectService.getOneByToken(projectToken);
         if (!projectOptional.isPresent()) {
             throw new ObjectNotFoundException(Project.class, projectToken);
+        }
+
+        Project project = projectOptional.get();
+        if (!projectService.isConnectedUserCanAccessToProject(project, authentication.getUserAuthentication())) {
+            throw new ApiException("The user is not allowed to modify this resource", ApiErrorEnum.NOT_AUTHORIZED);
+        }
+
+        Optional<User> userOptional = userService.getOne(userId);
+        if (!userOptional.isPresent()) {
+            throw new ObjectNotFoundException(User.class, userId);
         }
 
         projectService.deleteUserFromProject(userOptional.get(), projectOptional.get());
